+++ CHANGELOG.md
@@ -17,7 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * `unpack` structs
 * `column` index access
 * `toSql` and `sql` compiles and runs SQL query
-* array column scalars
+* array column scalars and filter
 * join and set fields
 * `implement` and `typed` for customized types and instances
 * rank and aggregate `over` grouping windows